Put all the stubs in a single file
