+++ Tests/APIClient/RouterTests.swift
@@ -1,0 +1,12 @@
+//
+//  Created by Pierluigi Cifani on 07/05/2018.
+//
+
+import XCTest
+@testable import BSWFoundation
+
+class RouterTests: XCTestCase {
+    func testURLEncoding() throws {
+
+    }
+}
